use nptyping for type hints only (#19)
The nptyping package isn't being actively maintained and is currently incompatible with numpy 2.0+. This package should be compatible with any reasonably recent numpy version. This release makes nptyping an optional dependency and uses it for type hints only. It also adds explicit tests for compatibility with all versions of numpy between 1.25.0 and 2.3.0.