Optionally parallelize Rust workload (#27)
The mergechannels workload is embarrassingly parallel. All input arrays are read-only and each index of the output array is created independently. This PR adds a parallel boolean argument which uses a rayon threadpool to colorize images when True. The overhead of the threadpool may be a net negative for small images (in the neighborhood of 256x256), but is typically a 2-9x performance improvement for normal (512x512) to large (2048x2048) images. The number of maximum threads in the rayon threadpool can be capped by setting the RAYON_NUM_THREADS environment variable if necessary. The argument parallel=True is now the default for the public facing API because it tends to be either a major improvement or already so fast that the threadpool overhead is negligible. Using the rayon threadpool can also offer an additional performance benefit with multiple python threads in a no-GIL python runtime. For sensitive applications you should run your own benchmarks and adjust accordingly.